requirements.txt
# Data generation pipeline (data_generation/)
openai>=0.28,<1.0
google-generativeai>=0.3.0
anthropic>=0.18.0
tqdm>=4.60.0

# Experiments (experiments/)
torch>=2.1.0
transformers>=4.43.0
accelerate>=0.25.0
numpy>=1.24.0
pandas>=2.0.0
